+++ packages/next-example/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.11.0](https://github.com/Portkey-Wallet/portkey-web/compare/v2.11.0-alpha.16...v2.11.0) (2024-10-08)
+
+## [2.10.6-alpha.14](https://github.com/Portkey-Wallet/portkey-web/compare/v2.10.6-alpha.13...v2.10.6-alpha.14) (2024-09-30)
+
+## [2.10.6-alpha.13](https://github.com/Portkey-Wallet/portkey-web/compare/v2.11.0-alpha.14...v2.10.6-alpha.13) (2024-09-30)
+
+## [2.10.6-alpha.12](https://github.com/Portkey-Wallet/portkey-web/compare/v2.11.0-alpha.13...v2.10.6-alpha.12) (2024-09-30)
+
+## [2.10.6-alpha.11](https://github.com/Portkey-Wallet/portkey-web/compare/v2.11.0-alpha.12...v2.10.6-alpha.11) (2024-09-30)
+
+### Bug Fixes
+
+- 🐛 multi transaction fix params ([b051c1d](https://github.com/Portkey-Wallet/portkey-web/commit/b051c1da9034c6d364b95c334f1c45e96269427f))
+
+## [2.10.6-alpha.10](https://github.com/Portkey-Wallet/portkey-web/compare/v2.11.0-alpha.10...v2.10.6-alpha.10) (2024-09-29)
+
+### Features
+
+- 🎸 multi transaction fix params ([7441905](https://github.com/Portkey-Wallet/portkey-web/commit/74419059d06cfffdf7d088cfe7fe5198c53c0534))
+
+## [2.10.6-alpha.9](https://github.com/Portkey-Wallet/portkey-web/compare/v2.11.0-alpha.6...v2.10.6-alpha.9) (2024-09-29)
+
+## [2.10.6-alpha.8](https://github.com/Portkey-Wallet/portkey-web/compare/v2.10.6-alpha.7...v2.10.6-alpha.8) (2024-09-27)
+
+## [2.10.6-alpha.7](https://github.com/Portkey-Wallet/portkey-web/compare/v2.10.6-alpha.6...v2.10.6-alpha.7) (2024-09-27)
+
+## [2.10.6-alpha.6](https://github.com/Portkey-Wallet/portkey-web/compare/v2.10.6-alpha.5...v2.10.6-alpha.6) (2024-09-27)
+
+## [2.10.6-alpha.4](https://github.com/Portkey-Wallet/portkey-web/compare/v2.10.6-alpha.3...v2.10.6-alpha.4) (2024-09-27)
+
+## [2.10.6-alpha.3](https://github.com/Portkey-Wallet/portkey-web/compare/v2.11.0-alpha.3...v2.10.6-alpha.3) (2024-09-27)
+
 ## [2.11.0-alpha.16](https://github.com/Portkey-Wallet/portkey-web/compare/v2.11.0-alpha.15...v2.11.0-alpha.16) (2024-09-30)
 
 **Note:** Version bump only for package @portkey/next-example